@@ -1,46 +1,34 @@
-import { IonApp, IonRouterOutlet, setupIonicReact } from "@ionic/react";
-import { IonReactRouter } from "@ionic/react-router";
-import type React from "react";
-import { Redirect, Route } from "react-router-dom";
 
-// Pages
+import { Toaster } from "@/components/ui/toaster";
+import { Toaster as Sonner } from "@/components/ui/sonner";
+import { TooltipProvider } from "@/components/ui/tooltip";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import HomePage from "./pages/HomePage";
-import ImageSearchPage from "./pages/ImageSearchPage";
-import SearchResultsPage from "./pages/SearchResultsPage";
-import TextSearchPage from "./pages/TextSearchPage";
+import NotFound from "./pages/NotFound";
+import LensSearch from "./pages/LensSearch";
+import LensResults from "./pages/LensResults";
+import { AppProvider } from "./context/AppContext";
 
-// Ionic CSS
-import "@ionic/react/css/core.css";
-import "@ionic/react/css/display.css";
-import "@ionic/react/css/flex-utils.css";
-import "@ionic/react/css/float-elements.css";
-import "@ionic/react/css/normalize.css";
-import "@ionic/react/css/padding.css";
-import "@ionic/react/css/structure.css";
-import "@ionic/react/css/text-alignment.css";
-import "@ionic/react/css/text-transformation.css";
-import "@ionic/react/css/typography.css";
+const queryClient = new QueryClient();
 
-// Theme and global styles
-import "./theme/global.css";
-import "./theme/variables.css";
-
-setupIonicReact();
-
-const App: React.FC = () => (
-  <IonApp>
-    <IonReactRouter>
-      <IonRouterOutlet>
-        <Route exact path="/home" component={HomePage} />
-        <Route exact path="/text-search" component={TextSearchPage} />
-        <Route exact path="/image-search" component={ImageSearchPage} />
-        <Route exact path="/search-results" component={SearchResultsPage} />
-        <Route exact path="/">
-          <Redirect to="/home" />
-        </Route>
-      </IonRouterOutlet>
-    </IonReactRouter>
-  </IonApp>
+const App = () => (
+  <QueryClientProvider client={queryClient}>
+    <TooltipProvider>
+      <AppProvider>
+        <Toaster />
+        <Sonner />
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<HomePage />} />
+            <Route path="/lens" element={<LensSearch />} />
+            <Route path="/lens/results" element={<LensResults />} />
+            <Route path="*" element={<NotFound />} />
+          </Routes>
+        </BrowserRouter>
+      </AppProvider>
+    </TooltipProvider>
+  </QueryClientProvider>
 );
 
 export default App;
